+++ admin/php/lxd/cluster.php
@@ -5,22 +5,12 @@ if (!empty($_SERVER['PHP_AUTH_USER'])) {
   //Instantiate the GET variables
   if (isset($_GET['remote']))
     $remote = filter_var(urldecode($_GET['remote']), FILTER_SANITIZE_STRING);
-  if (isset($_GET['instance']))  
-    $instance = filter_var(urldecode($_GET['instance']), FILTER_SANITIZE_STRING);
-  if (isset($_GET['project']))
-    $project = filter_var(urldecode($_GET['project']), FILTER_SANITIZE_STRING);
   if (isset($_GET['action']))
     $action = filter_var(urldecode($_GET['action']), FILTER_SANITIZE_STRING);
   if (isset($_GET['name']))
     $name = filter_var(urldecode($_GET['name']), FILTER_SANITIZE_STRING);
-  if (isset($_GET['profile']))
-    $profile = filter_var(urldecode($_GET['profile']), FILTER_SANITIZE_STRING);
-  if (isset($_GET['fingerprint']))
-    $fingerprint = filter_var(urldecode($_GET['fingerprint']), FILTER_SANITIZE_STRING);
-
-  //Instantiate the POST variable
-  if (isset($_POST['json']))  
-    $json = $_POST['json'];
+  if (isset($_GET['force']))
+    $force = filter_var(urldecode($_GET['force']), FILTER_SANITIZE_STRING);
 
   //Set the curl variables
   $cert = "/var/lxdware/data/lxd/client.crt";
@@ -35,25 +25,15 @@ if (!empty($_SERVER['PHP_AUTH_USER'])) {
 
     //Run the matching action
     switch ($action) {
-      case "deleteProject":
-        $url = $url . "/1.0/projects/" . $name;
+      case "deleteClusterMember":
+        if ($force == "true"){
+          $url = $url . "/1.0/cluster/members/" . $name . "?force=1";
+        }
+        else {
+          $url = $url . "/1.0/cluster/members/" . $name;
+        }
         $results = shell_exec("sudo curl -k -L --connect-timeout 3 --cert $cert --key $key -X DELETE $url");
         break;
-      case "createProject":
-        $url = $url . "/1.0/projects";
-        $data = escapeshellarg('{"name":"' . $name . '"}');
-        $results = shell_exec("sudo curl -k -L --connect-timeout 3 --cert $cert --key $key -X POST -d $data $url");
-      break;
-      case "loadProject":
-        $url = $url . "/1.0/projects/" . $name;
-        $results = shell_exec("sudo curl -k -L --connect-timeout 3 --cert $cert --key $key -X GET $url");
-      break;
-      case "updateProject":
-        $url = $url . "/1.0/projects/" . $project;
-        $data = $json;
-        $results = shell_exec("sudo curl -k -L --connect-timeout 3 --cert $cert --key $key -X PUT -d '$data' $url");
-      break;
-
     }
   }
 
@@ -63,5 +43,5 @@ if (!empty($_SERVER['PHP_AUTH_USER'])) {
 else {
   echo '{"error": "not authenticated", "error_code": "401", "metadata": {"err": "not authenticated", "status_code": "401"}}';
 }
-
+  
 ?>